Added int the networking final assignemnt
You can control players with wasd and shoot with spacebar.
Shoot the player with the flage to make them drop it and ragdoll
There is 2 powerups invincibility to help the flag carrier and increased fire rate for other players.
You can play games from 2-4 players and restart the game with the hosting player when a match ends.